After a great call with Jacob, moved some things around to avoid an edge case that could possibly affect us

+++ internal/config/concurrency.go
@@ -6,7 +6,7 @@ import (
 
 // ConfigureConcurrencyLimits configures the per-repo, per RPC rate limits
 func ConfigureConcurrencyLimits() {
-	maxConcurrencyPerRepoPerRPC := make(map[string]int64)
+	maxConcurrencyPerRepoPerRPC := make(map[string]int)
 
 	for _, v := range Config.Concurrency {
 		maxConcurrencyPerRepoPerRPC[v.RPC] = v.MaxPerRepo
